// src/components/form/form.interface.tsx
export interface IForm {
  diagnosed_with_problem: string;
  physical_trauma_start_after: string;
  mental_trauma_start_after: string;
  how_often_do_you_experi_problem: string;
  notRelevent: boolean;
  whenLyingDown: boolean;
  whenSitting: boolean;
  whenStanding: boolean;
  whenWalking: boolean;
  other: string;
  how_intensive_is_the_experience: string;
  id: number;
}